feat: add api

+++ garage_server/db/database/init.go
@@ -1,6 +1,6 @@
 package database
 
 type Pagination struct {
-	Limit  uint `validate:"max=100,min=1,number"`
-	Offset uint `validate:"min=0,number"`
+	Limit  uint64 `validate:"max=100,min=1,number"`
+	Offset uint64 `validate:"min=0,number"`
 }
