+++ garage_server/handler/anime.go
@@ -82,8 +82,8 @@ func (h *animeHandle) DeleteAnime(ctx *fiber.Ctx) error {
 // @Router       /anime [get]
 func (h *animeHandle) GetAnimes(ctx *fiber.Ctx) error {
 	var p = database.Pagination{
-		Limit:  uint(ctx.QueryInt("page_size", 50)),
-		Offset: uint(ctx.QueryInt("page_size", 50) * ctx.QueryInt("page", 0)),
+		Limit:  uint64(ctx.QueryInt("page_size", 50)),
+		Offset: uint64(ctx.QueryInt("page_size", 50) * ctx.QueryInt("page", 0)),
 	}
 
 	data, err := h.svc.GetAnimes(&p)
